third_party/blink/web_tests/fast/forms/select/select-set-length-with-mutation-reorder-expected.txt

Tests that setting the .length on an HTMLSelectElement works in the presence of DOM mutation listeners that reorder option elements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS sel.length is 4
PASS sel.length is 2
PASS sel.options.item(0).id is "a"
PASS sel.options.item(1).id is "b"
PASS successfullyParsed is true

TEST COMPLETE

a
b
